requirements.txt
numpy==1.16.4
pandas==0.24.2
git+https://github.com/h2oai/pandas-ml#egg=pandas-ml
pycryptodome==3.6.6
requests==2.20.0
setproctitle==1.1.10
statsmodels==0.9.0
filelock==3.0.12
numpy==1.16.4
pandas==0.24.2
git+https://github.com/h2oai/pandas-ml#egg=pandas-ml
requests==2.20.0
scikit-learn==0.21.2
scipy==1.2.2
setproctitle==1.1.10
toml==0.10.0
tensorflow==1.13.1
datatable-0.8.0.dai-1.73-cp36-cp36m-linux_x86_64.whl
h2o4gpu-0.3.2-cp36-cp36m-linux_x86_64.whl
h2oaicore-1.7.0-cp36-cp36m-linux_x86_64.whl
scoring_h2oai_experiment_4e86cfd8_ba96_11e9_b4c6_0242ac110002-1.0.0-py3-none-any.whl
